[Analysis API] Use builtin environment in tests which is closer to the production one
In production (IDE & Standalone modes), builtins usually reside in the
stdlib, which is bundled into the IDE/Standalone mode application. The
JVM stdlib used in the project being analyzed by the Analysis API is a
different stdlib.

Before the commit, VirtualFiles, from which we took the builtins in both
cases, were the same builtins. Now, they are different, and for stdlib
builtins, we now create VirtualFiles in the air. This allows us to
decompile our builtins in both cases differently. See KT-61757 for the context.

KT-68689