[FIR] Retrieve TA annotation constructors from the TA constructor scope
This is how it's normally done when instantiating classes referenced
via typealiases, otherwise we risk mismatching the type arguments:
when consuming the original constructor for MyAnnotation, we see that it
has 1 type argument, but the call site `@FixedAnnotation` specifies
none. Meanwhile, wrapper-constructors we create for typealiases have the
same number of arguments as the typealias itself, so they already handle
the issue.

This change in `CallAndReferenceGenerator` is backed by:
- `FirOldFrontendMPPDiagnosticsWithPsiTestGenerated.Hmpp.MultiplatformCompositeAnalysis.AnnotationMatching#testTypealiasedAnnotation`
- `FirOldFrontendMPPDiagnosticsWithPsiTestGenerated.ActualAnnotationsNotMatchExpect#testWithAnnotationActualTypealias`

There we get an error that some `IrConstructorSymbol`s are not bound.
This is because now these symbols are generated for wrapper constructors
for typealiases. Sinse we only need these wrapper constructors during
the resolution stages (for mapping arguments and whatnot), we can unwrap
them back for IR.

Since the typealias constructor is unwrapped manually, in FIR2Ir we need
to make sure the type arguments are manually remapped. Unfortunately, I
couldn't write a test that would fail due to this :( The problem can be
reproduced and verified via the debugger, though, with the help of the
following snippet and
`FirLightTreeJvmIntegrationDiagnosticTestGenerated.java`:

```
@Retention(AnnotationRetention.BINARY)
annotation class MyPlatformAnnotation<K, V>()

typealias Test<T> = MyPlatformAnnotation<T, T>

@Test<Int>
fun someFun() {}
```

^KT-68996 Fixed

(cherry picked from commit b25542cbb4f484bf920e592f44af3a1a3009cd7b)