[K/N] XCTest: Add state into the Firebase test runner
Running tests in Firebase requires an Xcode project to be built.
However, it may end up with Xcode or project configuration or
installation problems. Add state to be checked before
the execution to provide a fail fast strategy: don't try to build
and waste build execution time of something already went wrong.