[PowerAssert] Ignore implicit receivers of calls
To make sure the diagram only prints a single value at each offset,
ignore implicit receivers, which are currently displayed at the same
offset as the call itself. Only explicit receivers should be included to
avoid any confusion about which value is the receiver and which value is
the call result. If the display of the receiver is still desired, an
explicit `this` can be used.

^KT-69290 Fixed