K2: Fix controversial inference results with `!!` operator
The problem was that inference starts working differently depending on
the presence of non-trivial bound (Any instead of Any?), see KT-69266
and KT-69326.

The fix for the issues above would lead to re-work of how we choose the
order in which we fix the variables.

But the cases from the issue were working in K1 correctly because
synthetic call for `!!` has a different signature (not having an upper
bound for the type parameter).

Thus, the solution is just repeating K1 behavior here, too.

^KT-69159 Fixed