[KxSerialization] Made generatedSerializer function internal
By making the generatedSerializer() function, we share the implementation details.
The most common use case is the use of a generated serializer inside a custom one, which is declared in the same module as the serializable class.
For rare cases, the author of the serializable class can create a public function to share generated serializer.

Relates https://github.com/Kotlin/kotlinx.serialization/issues/1169


Merge-request: KT-MR-16602
Merged-by: Sergei Shanshin <Sergey.Shanshin@jetbrains.com>