[Gradle] Fix Configuration Cache for GMDT task
1. parentLibrariesIndexFiles should be a "SetProperty"
so that it makes it can be zipped latter.

2. In allTransformedLibraries ownRecords should be read
in a separate map. Attempt to write something like this
`parentLibrariesIndexFiles.zip(transformedLibrariesIndexFile) {...`
will fail configuration cache

3. RegularFile.records() must be a JVM static method, attempt to
extract it as Task method will cause this task to leak in "mapping"
lambda and cause configuration cache failure

Yes, configuration cache is hard and fragile.

^KT-69129 Verification Pending