[JS] Don't forget to visit IrReturns in SecondaryConstructorLowering
The reason why in this code:

```
constructor(void: String) : this({ x -> x + void })
```

the usage of `void` inside the lambda was not escaped is because
we forgot to run the value remapper on the function expression's
return statement, and instead left the return statement untouched.

This had the effect that the parameter's use site and its declaration
were not in any way linked in the IR, which confused the code generator,
which didn't know these two things refer to the same parameter.

Fix this by running the value remapper on the children of the return
statement as well.

^KT-66898 Fixed