JVM: make usages of indy lambda parameters nullable
The code in subsequent JvmInlineClassLowering is very sensitive of the
exact types of expressions. In particular, if it sees an equals
comparison where the left operand is boxed and the right operand is
unboxed, it unboxes the first operand and invokes an optimized static
equals from the inline class.

In the newly added test, in the `l2` case, it was incorrectly determined
that the right operand is unboxed. In fact it is boxed because
`LambdaMetafactoryArguments.adaptLambdaSignature` changed the parameter
type. But the usage was not updated, so the expression type was seen as
non-null.

 #KT-69282 Fixed