[K/N] Add bucketing of FixedBlockPage allocations
There is an overhead associated with having many size categories for
FixedBlockPage. This commit addresses this by rounding the allocation
size up before allocating. The specific strategy is keep the first three
significant bits of the requested size (measured in 8 byte blocks), and
round the rest up to 1-bits.  E.g., 0b1100101 would get rounded up to
0b1101111. This reduces the number of unique allocation sizes. Before
this commit, 128 unique types of FixedBlockPage was needed to cover
allocations up to 1kb. With this commit, only 23 is required for the
same range.