[PowerAssert] Capture safe cast expressions
Previously the only type operations captured by power-assert where type
checks. This caused safe casts to not be included in the diagram, and
more alarmingly, ignored if they were the root expression of a
parameter. This resulted in incorrect behavior of power-assert
transformed code in certain situations.

To fix, capture safe cast expressions and include all other type
operations as a constant expression.

^KT-69401 Fixed