[FIR] Smartcast implications when both sides of equality are literals
Within data flow analysis, we can rely more on the creation of a real
variable when add implications. Even if both sides of an equality check
are literals, adding implications may be required if the type of a
variable is explicitly defined to achieve smart casting. Instead of
making repeated checks to determine what sides are constants, if either
side is constant, attempt to apply implications and let real variable
creation reject processing as needed.

^KT-69392 Fixed