[Compose] [nonSkippingGroupOptimization] Fixes group eliding for if statements
With nonSkippingGroupOptimization enabled the rules of when a group can be elided
change. It is now required that groups generate a constant number of groups unless
they are specifically looping groups which can only generate the body of the loop
repeated, each iteration must be a single group. The groups for if statements were
still being elided not following this rule.

Fixes: [346821372](https://issuetracker.google.com/346821372)
Relnote: Fixes group generation for if statements when nonSkippingGroupOptimziation is enabled.