[stdlib] Mark DefaultConstructorMarker as @PublishedApi
When we generate a synthetic accessor for a private
constructor (KT-64865), we should be able to distinguish somehow
between the accessor constructor and the private constructor.

For that, we'll add a marker argument with type
`DefaultConstructorMarker?` to the accessor constuctor.

This is how it's already done in Kotlin/JVM.

Since we want to be able to serialize such accessors into KLIBs,
this `DefaultConstructorMarker` will become public ABI.

It should be marked as such.