FIR: always copy property accessors for delegated origin
And substitution overrides because they're encountered when the
interface that is being delegated to is generic.

Also, keep metadata for property accessors when changing origin of fake
overrides to delegation. This metadata is used in the backend when
generating JVM signatures in the metadata, so this fixes KT-68997 and
KT-63828. However, there's an edge case which is not resolved and should
be investigated separately in KT-69522.

 #KT-63828 Fixed
 #KT-63871 Fixed
 #KT-68997 Fixed
 #KT-69522

(cherry picked from commit e310c5eb02be6aacbdcd1e19cce031c6fd776803)