IR Inliner: Add test for references to inline functions that are not inlined
The inliner currently doesn't support inlining references to
inline functions. So, such references remain after the inliner
in the IR tree.

We need to teach the inliner to handle such references. Or to
put a dedicated lowering prior the inliner that will transform
such references into references to generated local functions
that do a normal call to inline function.

^KT-69172
^KT-69457