FIR: fix modality of property accessors in interfaces
For a simple abstract property in an interface, this code previously
calculated its default accessors' modality to be OPEN, which is
incorrect. It should be equal to the property modality (ABSTRACT)
instead.

At the moment it doesn't seem to affect anything, but a subsequent
change makes it so that FirFakeOverrideGenerator starts to generate
accessors for properties inherited by delegation, and having incorrect
modality there leads to differences in IR text tests in case of
delegation to Java interfaces.