Tests: change setter parameter name in IR text tests
After previous commit, setters of some properties inherited from Java
interfaces will now have the name `value` (default in K2) instead of
`<set-?>` (default in K1, and used in fir2ir), because they are now
generated in the frontend via FirFakeOverrideGenerator instead of in
fir2ir.

This change can be avoided by alowing to supply a custom setter
parameter name in FirFakeOverrideGenerator, but that seems like an
overkill because in practice, this name does not affect anything. Also,
if it needs to be fixed, it should be addressed systematically in
KT-63984.

(cherry picked from commit ba7a536aa46e5b18de9ef8c8421863552ee9a9c0)