[Tests] Update `PrioritizedEnumEntries` test-data (2.1)
The feature affects resolving that causes some red tests because of
ambiguity between newly `entries` property and explicitly declared `entries` by a user.
It was decided to disable the feature for the test with disabled `-EnumEntries`
because of the combination `-EnumEntries, +PrioritizedEnumEntries` doesn't make sense,
and `-EnumEntries` is outdated behavior, that it is a minor issue.
Probably some of those tests become redundant/useless, but it can be fixed later if needed.

KT-69560