Android-extensions: remove obsolete parcel part
Android-extensions is an obsolete plugin that cannot be used since
Kotlin 1.8.0 because an unconditional error is reported. The only reason
it is still not removed from the repo is because there are usages from
the Android part of intellij. In this change, the "parcelize" part of
the plugin and its tests are removed because it's not used from AS.