amend! [FIR] Add a test for `FirAccidentalOverrideClashChecker`
The first test is sensitive to the specified `MppCheckerKind.Platform`:
`Common` would erase `ACCIDENTAL_OVERRIDE_CLASH_BY_JVM_SIGNATURE`.
The diagnostic must be present, though, otherwise through the `expect`
class it would be possible to declare something that is incorrect
from the PoV of the `actual` class (can't be declared there).

The second test represents a situation when something that we can't
write in an `actual class`, because it would trigger a clash, can still
be written in the corresponding `expect` class. But this case will be
caught by an `IrActualizationError`.

^KT-68000