[Native] `isFromCInteropLibrary()`: Don't process IR based descriptors
The goal of `isFromCInteropLibrary()` function is to check whether
the given declaration is from a C-interop library. For such
declarations either of the following conditions should be met:

1. It is Fir2Ir lazy declaration built around the corresponding
   FIR declaration. And we can look into the FIR declaration to
   figure out it's <source>.

2. It is an IR declaration built around the deserialized metadata (K1).
   In different compiler modes it can be either an instance of
   `IrLazyDeclarationBase`, or an instance of `IrDeclaration`, but
   always with the appropriate DeclarationDescriptor.

That means the following: If we meet an IR declaration that is not
Fir2Ir lazy declaration and has no real descriptor, then it can't
be a declaration deserialized from a C-interop library.

^KT-69563