Make inference work properly with lambdas in return statements of PCLA
It's only related to PCLA lambdas with return types containing not-fixed
type variables

Previously, they were being analyzed outside the PCLA inference session
that led to inconsistencies in the CS.
It happened because they were being started inside a common CS, which
 was being committed back to the CS of the main candidate after PCLA.

Initially, the problem appeared with KT-68940, which was quick-fixed
with assertion muting.

But the exception from KT-69040 was still relevant.

One of the fixes might be somehow re-animating inference session and
common CS when returning to the nested lambda analysis. However, that
would lead to the controversial semantics because nested lambda
might be analyzed after the fixation of the builder type variable
to a too restrictive type.
See test2 in lambdaBelongsToOuterCall.kt: before running analysis of
`c.foo {` lambda we would fix `BKv` to `A` which would lead to type
mismatch in `c.add(b)`.

For more details, see the changes in docs/fir/pcla.md

^KT-69040 Fixed