IR: do not remap types for lambdas extracted to local functions
Sometimes extracted local function should really reference the original
function's type parameter, such as in KT-68675 where it happens for a
type argument passed for a reified type parameter. Remapping it to the
local function's type parameter is problematic for at least two reasons:

1) Semantically, `typeOf` references the original function's type
   parameter, and not the one of the moved function (which usually has
   a name like `g$lambda$0` and is synthetic)
2) On JVM, we do not generate type parameters of extracted local
   functions at all for other reasons (see `FunctionCodegen.kt:62`), so
   kotlin-reflect cannot find such function in the class.

This problem has appeared in 2.0 because indy lambdas have been enabled
by default. In `-Xlambdas=class`, the code from KT-68675 worked because
before LocalDeclarationsLowering, we transformed lambda into an
anonymous class with invoke, and everything worked fine because the type
parameter was not remapped. (Local classes are not extracted to top
level on JVM in most cases -- only in property/anonymous initializers,
which can't be generic.)

 #KT-68675 Fixed
 #KT-69433 Fixed