[FIR] Enhance boolean logic DFA to support all boolean literals
After the language feature InferMoreImplicationsFromBooleanExpressions,
some boolean literals in boolean logic are correctly understood. But it
is incomplete which leaves DFA with inconsistencies. Make sure that
`false` on either side of an or operation, and `true` on either side of
an and operation, are both correctly understood as being useless to the
the final result of the boolean logic.

^KT-69537 Fixed