[FIR] Explicitly set left and right for binary logic exit CFG node
Instead of calculating left and right operand nodes each time they are
accessed from the list of previous nodes, construct the binary exit node
with the left and right nodes. They are already known when constructing
the exit node, so this is not only a less brittle way to access these
nodes, but possibly a small performance improvement.

^KT-59684