[K/N][Tests] Fix `TestRunParameter.WithGTestPatterns` in the test infra
`WithGTestPatterns` test parameter doesn't have any effect at the
moment: because of a typo introduced in
2276abfb85d0baa944c03a3234f0ae4e84a5a1ca, it didn't add the
corresponding argument to the test runner. As a result, ignoring tests
with this parameter didn't work.

This commit fixes the problem by fixing the typo. Now the argument gets
added to the test runner, and ignoring the tests works as expected.