[K/N][Tests] Fix MacOSLinkerIncludedUniversalBinariesTest on devices
MacOSLinkerIncludedUniversalBinariesTest test checks that linker can
handle universal binaries properly. For that it compiles two binaries
with `clang` for different architectures -- arm64 and x86_64, and
combines them with lipo.

The problem is that compiling to x86_64 doesn't work for when targeting
real devices (e.g. ios_arm64), because this arch is not supported by the
corresponding SDKs.

This commit fixes the problem by compiling for arm64e instead of x86_64
in such cases.