[PowerAssert] Fix indenting of power-assert diagram when tabs are used
Power-assert currently assumes that spaces are used for code indents by
prepending diagram display lines with the number of spaces as the column
offset of the expression. However, if tabs are used for indents, this
causes expression result display to be misaligned. Instead, create a
template of the source code line by replacing non-white space with
spaces, and use substrings of that to prepend diagram display lines.
This preserves tabs which are present within the source code.

^KT-70113 Fixed