[JVM] Remove line numbers from parameter destructuring in lambdas
For the list of debugger issues that are solved by this commit see
IDEA-335314. This commit also changes outputs of some tests. I will break
these changes down below.

# JDB stop on method entry
org.jetbrains.kotlin.test.runners.codegen.inlineScopes.FirLocalVariableIrInlinerTestWithInlineScopesGenerated#testUnderscoreNames
org.jetbrains.kotlin.test.runners.codegen.inlineScopes.FirLocalVariableIrInlinerTestWithInlineScopesGenerated.Destructuring#testLambdaMultiline
org.jetbrains.kotlin.test.runners.codegen.inlineScopes.FirLocalVariableIrInlinerTestWithInlineScopesGenerated.Destructuring#testLambdaCustomComponentNsMultiline

When the VM steps into a method, it will stop in the beginning of it,
even if the first line number is declared a number of opcodes later.
When stepping into a lambda with parameter destructuring you will stop not on
the first line number declared in it, but on the method entry instead, and the
parameters won't be visible in the debugger as they haven't been destructured
yet. This inconvinience can only be met when debugging outside of Intellij
IDEA, for example with a command line JDB, as the Intellij debugger deals with
this issue by doing one opcode wise step to reach the first declared line
number in a lambda.

# SMAP is not generated because it is not needed anymore
org.jetbrains.kotlin.test.runners.JvmAbiConsistencyTestBoxGenerated.Contracts#testLambdaParameter

Consider this expression:
```
("O" to "").let { (a, _) ->
	...
}
```
As line numbers are not added for parameter destructuring, SMAP is no longer
needed. The first location declared in a lambda will point to the body of it
with a known line number.

# IR difference
org.jetbrains.kotlin.test.runners.ir.FirLightTreeJvmIrTextTestGenerated.Lambdas#testDestructuringInLambda

The commit adds a new variable origin which has to be reflected in tests.