[Parcelize] prioritize `@DataClass` over `java.io.Serializable`
Supporting both at the same time requires some refactoring of the
plugin, as a data class may fail the requirements on visibility or
contained types but still implement `Serializable`. So for now, if a
type is annotated with `@DataClass`, even `Serializable` data classes
inside it will use the new logic. So this will emit an error:

    data class C(val x: Any) : Serializable

    @Parcelize
    class D(val c: @DataClass C) : Parcelable

Note that this does not affect anything not annotated `@DataClass`, and
`Serializable` non-data class members are allowed as well. So this is
fine:

    data class C(val x: Any) : Serializable

    @Parcelize
    class D(val c: C) : Parcelable // uses writeSerializable

And so is this:

    data class C(val x: Serializable)

    @Parcelize
    class D(val c: @DataClass C) : Parcelable