[DLC] do not throw exception in case of inconsistency by default
For some reason, it is rarely possible to have an inconsistency
between decompiled regular class and decompiled light class,
which leads to a flaky exception.
As the root cause is unknown, and we do not expect to see effects
in the production code, it makes sense to disable the exception
by default until we have more or less stable reproducer with real
use cases.

^KTIJ-24782 Fixed