[FIR] Rename `MppCheckerKind` to `CheckerSessionKind`
And `MppCheckerKind.Common` to
`CheckerSessionKind.DeclarationSiteForExpectsPlatformForOthers`.
This is done to avoid ambiguity that the word "common" brings when
talking about either the session kind or whether the checker is a `Jvm`/
`Js`/`Native` checker or not. The word "platform" is still ambiguous,
but it was decided to leave it as is.

Also, `DeclarationSiteForExpectsPlatformForOthers` makes it easier to
see how checkers are supposed to work in case of HMPP.

Note that this commit only renames the enum, but doesn't change the
behavior.