[Gradle KMP IT] Use the version of K/N produced earlier in the same build
For Gradle KMP tests, it is necessary to use the version of
Kotlin/Native that is aligned with other compilers:
Kotlin/JVM and Kotlin/JS. Practically, this means that the tests should
use the Kotlin/Native distribution produced in the same build with
other Kotlin compilers and not try to download some predefined
(and most often a bit outdated) Kotlin/Native distribution.

Why is it important?

We are going to bump KLIB ABI version. After the bump, the compiler
will start writing a new ABI version to each KLIB manifest file.
This will immediately take effect for the Kotlin/JS and metadata
compilers. However, an outdated version of the Kotlin/Native compiler
that is downloaded by the KGP will still be unaware of the ABI version
bump. At least for a period of time until we advance the version
of bundled Kotlin/Native in KGP. And during this time we would get
numerous "KLIB has incompatible ABI version" errors in KMP tests.

^KT-68792