[K2 Scripting] Use the script session for default import resolution
In the 'IGNORE_SELF' copy analysis mode in LL FIR, the designation
may contain declarations from the contextual module (while the target is
from the file copy).

The dangling file module's 'FirProvider' does not return files from
different modules, as expected. In the fix, we always use the session
owning the script declaration, instead of the use-site one.

As the test setup needed to reproduce the issue is quite sophisticated,
the test can be found on the IntelliJ Kotlin plugin side.

^KT-69360 Fixed