[Gradle] Add archivesTaskOutputAsFriendModule flag
This flag should disable redundant logic in Friends Artifacts Reolver
that adds Kotlin Target's archive file as a friend path to associated
compialtions. It should be enough to just use compiler outputs.
But this change requires proper migration.

Enable this flag for kotlin-stdlib as it's associated compilations
contribute to archives task. This should break cyclic dependency.

^KT-58280