[K/N][test] Do fflush to attempt to fix CExport testSimple on Windows CI
The test sometimes fails on Windows CI, because the output doesn't match
the expected value: the output contains only a few first lines.
It is possible that we need some `fflush` to make sure the output
is actually delivered. The test tried to do that, but the execution
actually goes a different way -- through the Kotlin unhandled exception
handler.

This commit adds `fflush(NULL)` there too in an attempt to fix the test
flakiness on Windows.

^KT-69308