[K/N] Do consoleFlush when terminating with an unhandled exception
This is an attempt to fix testUnhandledExceptionThroughBridge on
Windows. Sometimes the test fails on the CI, because the output is
missing the reported unhandled exception.

This commit adds `consoleFlush` (basically `fflush(stdout)` and
`fflush(stderr)`) to the code that reports an unhandled Kotlin exception
and exits abnormally. Flushing is added right before `std::abort()`.

^KT-69663