[K/N] Mute CExport testConcurrentTerminate on Mac
This test flaks on Mac CI, and it is not clear why. All failures happen
because the termination handler can't finish in 5 seconds, so the
failsafe kicks in and exits the process.
See `timeoutSec` in `Exceptions.cpp`.

This might have something to do with the test timeouts on Mac CI we've
been observing lately, so it doesn't necessarily indicate a bug.

^KT-69303