[K/N] Debug output for flaky testReleaseCompilerAgainstPreReleaseLibrary
The test fails on the CI from time to time. It seems that the compiler
prints full file paths instead of relative ones. This could happen if
something specifies a message renderer through system properties.

This commit adds logging for this message renderer system property value
in the failing test.

^KT-64822