[K/N] Update CallsChecker exclude list with `NSObject.isSubclassOfClass`
CallsChecker has a list of external LLVM function names that are safe
to call in the "Runnable" thread state.
197c989d75b04eb08d000dcb9a1d81ed327d8eee adds a call of
`isSubclassOfClass` Objective-C method in the Runnable state. It is
safe, so it needs to be added to the exclude list to make the tests
with CallsChecker enabled pass.
I.e. with `-Pkotlin.internal.native.test.useThreadStateChecker=true`.

This commit does that addition.

^KT-69469