[FIR] Ditch `KtTypeReference` from `FirDiagnosticsList`
Plugins can generate `FirTypeRef`s. There's no reliable way to provide
a `KtTypeReference` for them, so most of the time they will contain
incompatible ones. This is true for everything that can be generated
from a plugin, but until now we've only really faced it with type refs.

In the context of `entityTypeSuperClass`, it's important for the
`FINAL_SUPERTYPE` and `MANY_CLASSES_IN_SUPERTYPE_LIST` diagnostics,
otherwise we crash on them.

^KT-68375