[K/N][test] Update tests expecting to exceed default timeout
627cf0c9578164f63d36dc62b3274272bc67bf61 updates the default test
execution timeout to 60 seconds. But there are some tests (checking the
test infra itself) that wait for approximately 60 seconds and expect to
exceed the timeout (to check that timeout enforcing machinery actually
works). So now these tests started to fail randomly.

This commit adjusts those tests to wait 240 seconds instead of 60.
It also adds comments to remember to do this the next time.