IR Inliner: Inline `stub_for_inlining` at the right phase of double-inlining
For a function reference in a position of an inline function argument,
the `CommonInlineCallableReferenceToLambdaPhase` lowering generates
a lambda. The function expression in this lambda has an inline function
with the local visibility.

With the double inlining enabled, there is not enough information for
the inliner to deduce whether such a local inline function should
be inlined at the first or the second phase.

Using the visibility of the container inline function helps to solve
this problem.

^KT-69470