[FIR] Change the way `DeclarationSiteForExpectsPlatformForOthers` works
Previously, we'd have `MppCheckerKind.Common` to denote checkers that
are executed with the declaration-site session. Reviewing the usages
of this kind and thinking about it again we realised it's only used
to workaround expect declarations, producing boilerplate checker pairs
and confusing checker writers about the wrong default.

The new `CheckerSessionKind.DeclarationSiteForExpectsPlatformForActuals`
is a value specifically for that use case: if the checker is executed
for a declaration residing inside an `expect` declaration, then it's
run with the declaration-site session, otherwise it still uses the leaf
platform one. This change alters the behavior so that this enum entry
now denotes exactly what it means resulting in all the checkers marked
with it in working differently. Some of those checkers are pairs of
`Regular` and `ForExpectClass` for which nothing has changed, and some
are non-pair previously `Common` checkers that no behave "almost like"
`Platform` ones. Such checkers will later be reviewed and either
converted to `Platform` or left exactly `DeclarationSite`. The change
in behavior affects some tests, see the subsequent fixups.

Also note that there's a nuance in how paired checkers work (both before
this commit and after): previously metadata compilation would only run
`Platform` checkers. Meaning, in paired checkers, `Regular`s would still
be executed during metadata, they would receive the declaration-site
session, and they would also skip `expect` classes. This logic is
preserved in this commit by having a `null` value for the
declaration-site session holder.

Also, the `DeclarationSite` kind is added as well. It's basically the
old `Common`, but is now protected by an OptIn. It's not used in this
commit, but it will be assigned to some checkers in the later commits.
This way allows to get some failing tests (and there're not many of
them), and revert some checkers back to the old behavior if needed.

^KT-69561