Add the possibility to disable fast jar fs in K2
Note that it also enables fast jar FS for cases when LV >= 2.0, but
USE_FIR is not set to true.

While in general compiler setup for K2 we explicitly set USE_FIR=true,
the behavior for some tests is expected to be changed a bit.

But that kind of change (enabling FastJarFs when LV >= 2.0) looks fair
because that's how it actually works in the production compiler.

^KT-69792 Fixed