[FIR] Make `FirSealedSupertypeChecker` `DeclarationSite`
This checker compares `moduleData` of different declarations. To be
frank, I'm not sure I understand why this checker should work the way it
does, but I couldn't replicate this behavior with other
`CheckerSessionKind`s and satisfy the tests, and this is how it works in
K1, so let it be.