KT-69805: YarnSetupTask work with both custom and default downloads
Specifying a custom `downloadBaseUrl` for Yarn causes a `NullPointerException` as the file content is not extracted to match the destination field. 

This works for the default due to the downloaded tar archive name having a different name than the top-level directory in the tar, meaning the removePrefix actually did nothing. Remove altogether. 

More detailed explanation in: https://youtrack.jetbrains.com/issue/KT-69805/YarnSetupTask-does-not-work-for-custom-downloadBaseUrl