IR Inliner: New phase for dumping/testing synthetic accessors
Add a new `DumpSyntheticAccessorsPhase` phase that allows dumping
synthetic accessors and their use-sites. This is necessary for
tests on synthetic accessors in KLIB-based backends.

Limitations: Currently, synthetic accessors are generated inside
the lowerings pipeline in Native and JS backends. So, it's more natural
to add a new phase immediately after the generation of accessors
to dump them. If dumping was performed later (after the lowerings
pipeline -- after the backend phase), the IR would be significantly
lowered and not so appropriate for testing.

Later, when the synthetic accessors generation lowering would be
executed before KLIB serialization, we could consider converting
this phase into a regular `AbstractIrHandler` such as, for example,
`IrTextDumpHandler`.

^KT-69462