[SLC] do not trigger STATUS resolution there it is not needed
`STATUS` phase is required only in the case of value classes unwrapping.
In the light of it, there is no need to involve redundant resolution for
light classes as they are not supposed to work with value classes in
signatures.

Exceptional cases there this resolution is necessary:
- Top-level callables return type. We materialize such declarations
as they don't have mangling and are accessible from Java
- Backing fields. We materialize all of them
- Callables with `JvmName`. We materialize such declarations
as they don't have mangled names

Also, this commit fixes a possible contract violation for `TYPES` phase.

The scenario:

1. Someone wants to resolve function types, so it triggers TYPES phase
2. The resolver realizes that the requested type is Java library class
3. `JavaSymbolProvider` triggers `convertJavaClassToFir` to build a FIR
representation
4. To build the class, we have to build the class methods as well
5. To build a method, we need to know parameter names, and here starts
the tricky part
6. `JavaValueParameterImpl.getName` may decide to receive the name from
the underlying psi which is `ClsParameterImpl`
7.  To calculate the name, `ClsParameterImpl` tring to find the
corresponding method in library sources (`getSourceMirrorMethod`)
8. To match the method, Java resolved have to resolve parameter types
9. At this point, the resolver tries to iterate over all super types
10. One of its super types is Kotlin class. But the problem is that
this is a source class and not the binary one. It is possible because
the library may be built in a bootstrap way, so some interface may be
extracted into a library, but all super types are inside sources
11. As it is a source class, it is `SymbolLightClassForClassLike` on
which `getSupers` is called
12. To build this list (`createInheritanceList`), we need to map Kotlin
super types (`KtType`) into `PsiType`. Types itself are already resolved
as super types resolves during the `SUPER_TYPES` phase. But the previous
implementations of `asPsiType` triggered the `STATUS` phase resolution,
which is the contract violation as we started from the `TYPES` phase

^KT-69671 Fixed