Kapt: use IR instead of PSI to compute line information
This will help in implementing Kapt3 over K2 (KT-69860), where PSI is
not always available.

Line information is used to report errors and to sort declarations in
kapt. Ordering of some declarations was changed, but it's fine because
those are mostly synthetic declarations (data class-generated members,
bridges, etc.), which are unlikely to affect any real-world annotation
processors.

 #KT-69861 Fixed