[KT-69845][Compose] Fix the setting of the annotations of the IrClass instead of the annotations of the IrFunctionBuilder
AbstractComposeLowering.kt:947 sets the annotations of the extension
receiver IrClass instead of the IrFunctionBuilder.

This has resulted in some failure cases, like having Compose with Wasm
example that is using the `@Serializable` annotation fail because this
line was removing the `@Serializable` annotation when Compose is run
first.

Failure examples are attached in https://youtrack.jetbrains.com/issue/KT-69845