[Kotlin for IDE build] deprecate K1 resolve packages
We build kotlin-compiler-fe10-for-ide.jar, which contains Kotlin K1 compiler internals.
Declarations from this jar are used by 3rd party plugins in IntelliJ.
The K1 API is deprecated in favor of the Analysis API.

This commit marks the K1 API as deprecated by adding the package-info.java deprecation file to them.
This is done by generating a bunch of such files for each package we want to deprecate and putting them into the kotlin-compiler-fe10-for-ide.jar, which is a fat jar.

^KT-65469 fixed

(cherry picked from commit de5bb77dbbad2b5eb63320348dc9b728484a986b)