[FIR] Use `FirSession` as a part of `ScopeSessionKey` for `PACKAGE_MEMBER`
In some cases (HMPP), a common scope could leak to a platform scope.
For instance, `FirDesignatedCompilerRequiredAnnotationsResolveTransformer`
uses a session from a specified symbol (designation), but not the current one.
It caused bugs like unresolved members of `Throwable` during building stdlib with K2 (KT-69599).

Using `FirSession` as a part of `ScopeSessionKey` key
forces `scopeSession` to build different scopes for different sessions.