Fir2Ir: support error annotations, skip error ::class literals
This is needed for the re-implementation of K2 kapt. In kapt mode, error
types are allowed and should be somehow processed by fir2ir in a way
that allows the kapt plugin to recover them when generating Java stubs.
This fix affects the test `kt32596`, and fixes
`importsWithUnresolvedAnnotations` in K2 kapt.

 #KT-69860