[Gradle] Changed the way of getting psm data in GMT algorithm
The previous algorithm of getting psm data affected Gradle project-isolation.
That's why we moved psm info to consumable configurations for each multiplatform subproject.
Then, on the multiplatform consumer side, we have a resolvable configuration, which consumes the previous one.
The whole project isolation feature is covered with a new Feature Flag.
This flag is disabled by default.

^KT-64998 In Progress

Co-authored-by: Alexander Likhachev <Alexander.Likhachev@jetbrains.com>

Merge-request: KT-MR-16179
Merged-by: Dmitrii Krasnov <Dmitrii.Krasnov@jetbrains.com>