[K/N][test] Set test tags from the property
This property is being used on the TeamCity to separate subsets
of tests. The test task should use it if the tag wasn't set,
but the property was set by the TeamCity job.
Otherwise, test results may be inconsistent to the job descriptions,
like K2 tests being run for the job made only for K1 tests.