[FIR] Improve performance when tracking local variable assignment
During data-flow analysis, mutable local variables need not have their
assignments tracked before they can be smart-cast. This is in case a
local variable is mutated within a loop or captured and mutated within a
closure. There was a significant performance problem with this analysis
which could cause exponential degredation based on the number of when
branches or catch blocks.

^KT-69723 Fixed