[Native, Wasm, stdlib] Fix bug in quickSort for primitive arrays
^KT-70005 Fixed

It took 7 years, 3 months, 13 days; 3 project managers; 2 developers;
2 language designers; 2 fixup commits; 1 code reviewer; 1 sedative pill
to fix this bug

Jokes aside, the alternative fixes are:
         private fun quickSort(
                 array: LongArray, left: Int, right: Int) {
        +    if (left >= right) return
             val index = partition(array, left, right)
             if (left < index - 1)
                 quickSort(array, left, index - 1)
    OR:
         private fun quickSort(
                 array: LongArray, left: Int, right: Int) {
        +    if (left >= right) return
             val index = partition(array, left, right)
        -    if (left < index - 1)
        -        quickSort(array, left, index - 1)
        -    if (index < right)
        -        quickSort(array, index, right)
        +    quickSort(array, left, index - 1)
        +    quickSort(array, index, right)
         }

The first fix is bad because `left >= right` will never be `true` inside
the recursion. (quickSort is recursively invoked only when
`left < index - 1`) So we are doing a redundant check on every level of
the recursion

The second fix doesn't have the redundant boolean check, but in the
worst case it will increase the recursion by one, which might be just
enough to cause stack overflow and break code that worked before. Not to
mention, that in the recursion tree every recursion will take one more
recursion call to complete.

Co-authored-by: Igor Yakovlev <igor.yakovlev@jetbrains.com>
Co-authored-by: marat.akhin <marat.akhin@jetbrains.com>
Co-authored-by: Mikhail Zarechenskiy <mikhail.zarechenskiy@jetbrains.com>
Reviewed by: Ilya Gorbunov <ilya.gorbunov@jetbrains.com>