[AA] Add tests for several cases when type arguments size is not equal
...to type parameter size.

`doubleColonReceiverType` tests are to demonstrate that
`MyInterface::sam` will return type without args (and MyInterface<*>
in case of ::class).

`superTypes/callWithMissingTypeArgument` - to demonstrate the behavior,
works as expected
`superTypes/objectWithTypeParameter` - the only case from the issue
which it is possible to easily reproduce in tests.

Tests for all other cases from the issue will be added from the IDE
side.

^KT-65803