Kapt: do not use binding context to check class name
This will be needed for re-implementation of K2 kapt in KT-69860, where
binding context is empty. The intention of this code was to filter out
classes such as interface's DefaultImpls which shares origin with the
interface. But type mapping is not necessary here, simple source-based
internal name computation is sufficient.

 #KT-69860