IR Inliner: Support generating synthetic accessors in two modes
Mode 1 (with narrowing visibility): Narrow the visibility of a generated
accessor from _public_ to _internal_ if the accessor is only used in
_internal_ inline functions and therefore is not a part of public ABI.
To be used only at the first phase of compilation (before serializing
KLIB).

Mode 2 (without narrowing visibility): Always generate accessors with
the _public_ visibility. To be used only at the second phase of
compilation (in the backend).

^KT-69527