[K/N][test] Add value to EXPECTED_TIMEOUT_FAILURE test directive
There are some tests that expect to exceed the default timeout. The
problem is that we need to update the default timeout from time to time,
so these tests need to be updated as well. Moreover, the next step is to
make the default timeout very big (to help with the obscure timeout
failures on the CI). This would make the tests expecting to exceed
timeout take too much time unnecessarily.

Instead, this commit changes the EXPECTED_TIMEOUT_FAILURE test directive
so that it specifies the timeout value explicitly. This value is used
in the tests then, instead of the default one.

^KT-69784