[FIR] Introduce `-Xuse-fir-experimental-checkers`
Extended checkers were brought in as support for IDE inspections.
Therefore, they may be:

- inaccurate - have false positives
- inefficient - IDE decides which code to run checkers for, so it
  "balances" them so that this isn't observable to the user
    - in CLI the compiler compiles everything, so any slowness in any
      place is observable
- feels like a compiler warning - warns the user that their code may not
  work as they think it does (warns about semantics). Non-examples:
    - code-style specific
    - support for stdlib helpers

The users want to be able to run extended checkers themselves. Why?
Which of the above properties are important to them? We don't know, to
be frank. But we still want to have a buffer place we could put checkers
that are half-baked or not ready for publication.

What does half-baked mean? I propose to define half-baked as either:

- has false positives - in the presence of `-Werror` facing false
  positives would be a bummer
  - when someone uses `-Werror`, they agree to torture themselves, it's
    their responsibility, but they only agree to torture themselves with
    actual small problems, not false problems that we, in fact, know
    about
- has obviously slower-than-necessery implementation
  - the checker sometimes does some work twice for no reason, and
    somehow it got through the review process

So:

- If a checker is not half-baked & semantic -> default
- If a checker is not half-baked, but not semantic -> extra
- If a checker is half-baked -> experimental

Or if you are an automated prover: `half-baked -> experimental` and
`!semantic -> !default`.

`FirReturnsImpliesAnalyzer` is considered experimental because ^KT-68302
is open.

^KT-69939 Fixed
^KT-68302