[SLC] drop AnnotationUseSiteTargetFilter
This commit fixes the contract violation for the `TYPES` phase.
Previously, `hasAnnotation` for `JvmRecord` `ClassId` triggered
`KaAnnotationList.get` instead of `contains` since we have to
check `useSiteTargetFilter.isAllowed`.
Now this `hasAnnotation` will trigger only the `COMPILER_REQUIRED_ANNOTATIONS`
phase, so there are no contract violations.
As the next step, `get` should be improved as well to avoid the `TYPES`
phase request as we have lazy annotations.

^KT-70323 Fixed