[FIR2IR] Accept caching generated IR function for K2 CodeGen API
When we access `Fir2IrLazyClass::declarations` as a part of K2 CodeGen
API like an access to it from compiler plugin IR generation extension,
It creates some IR functions in Fir2IrDeclarationStorage and caches
them. `cacheIrFunctionSymbol()` accepts when the generated function is
`componentN()` or `copy()`, but cause an exception for `equals`,
`toString`, or `hashCode`. When it is the CLI compile, it is OK to
accept only `componentN()` or `copy()`, but K2 CodeGen API can have an
exception. This commit lets `cacheIrFunctionSymbol()` accept all
generated functions when it is called by K2 CodeGen API.

^KT-70351 Fixed
^KT-70376 Fixed