[K/N][caches] Support explicit stdlib passing
The cache directory for stdlib and platform libraries is selected differently as
opposed to all other libraries. This commit fixes a problem in this mechanism if
stdlib is passed explicitly (ot it would be confused with external libraries). Due to
the problem, the stdlib's cache was placed into the directory with other external
libraries caches, but the selecting cache mechanism in the compiler (in CachedLibraries.kt)
only searches for caches in this directory for external libraries (and stdlib isn't
considered external there because of its path).

 #KT-69835 Fixed