[IR] Don't load IR of lazy functions for anything but the body
in -Xserialize-ir mode. This is to be able to make them non-lazy, see
the next two commits.

In K1, those properties (mostly IrValueParameter) will be replaced by
their deserialized version at the time the body is deserialized.
In K2 it already works this way.

Which is generally bad, but should be OK enough, because there shouldn't
be any references to them outside of the body itself.
In fact this change seems to accidentally fix two tests of
-Xserialize-ir mode.