[Analysis API] KaAnnotation: deprecate `index`
The initial purpose of this property was to support lazy
annotations in symbol light classes, but now this logic is redundant
as we have lazy arguments.
This API doesn't have real value for users.
Also, currently it may not work correctly for special annotations as
the final annotation index is known only after the TYPES resolution
phase.

^KT-70323