[Analysis API FIR] AllCandidatesResolver: do not modify original code
The problem with `runCompletionForCall` is that it modifies lambdas
right away in-place. In the case of AllCandidatesResolver, it is
an original lambda from the resolved code. Hense, this is not safe
and broke everything. The modification logic is inside
`FirCallCompleter.LambdaAnalyzerImpl.analyzeAndGetLambdaReturnArguments`.
As a simple solution, we may use `ConstraintSystemCompletionMode.UNTIL_FIRST_LAMBDA`.
In this mode, the candidate will have the resolved type only in the case
of call without lambdas.
Cons: we cannot provide type parameters mapping in the case of ambiguity
errors case for `resolveToCallCandidates` API.

^KT-70199 Fixed